requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
flask-socketio==5.3.6
python-socketio==5.10.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
twilio==8.10.0
livekit==0.15.0
livekit-agents==0.8.0
deepgram-sdk==3.2.0
elevenlabs==1.1.2
openai==1.3.0
aiohttp==3.9.1
requests==2.31.0
gunicorn==21.2.0
